requirements.txt
fastapi 
uvicorn[standard] 
 
# Pydantic 
pydantic 
pydantic-settings 
 
# LangChain + OpenAI 
langchain 
langchain-openai 
langchain-community 
langchain-ollama==0.3.10
langchain-core==0.3.80
anyio
 
# Vector store 
chromadb 
 
# Database 
sqlalchemy 
psycopg2-binary 
 
# Env 
python-dotenv 
 
# PDF / docs 
pymupdf 
unstructured

# Plotting / visualization
matplotlib

# Data processing
pandas
openpyxl
